rename: add packages and move files of category
