+++ src/main/java/gift/api/category/dto/CategoryRequest.java
@@ -1,6 +1,6 @@
 package gift.api.category.dto;
 
-import gift.api.category.Category;
+import gift.api.category.domain.Category;
 
 public record CategoryRequest(
    String name,
