+++ src/main/java/gift/api/category/dto/CategoryResponse.java
@@ -1,4 +1,6 @@
-package gift.api.category;
+package gift.api.category.dto;
+
+import gift.api.category.domain.Category;
 
 public record CategoryResponse(
     Long id,
