+++ src/main/java/gift/api/member/controller/OauthController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.view.RedirectView;
 
 @Controller
-@RequestMapping("/oauth")
+@RequestMapping("/api/oauth")
 public class OauthController {
 
     private final KakaoProperties kakaoProperties;
