+++ src/main/java/gift/api/product/dto/ProductResponse.java
@@ -1,6 +1,6 @@
 package gift.api.product.dto;
 
-import gift.api.category.Category;
+import gift.api.category.domain.Category;
 import gift.api.product.domain.Product;
 
 public record ProductResponse(
