+++ src/test/java/gift/api/product/ProductTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.BDDMockito.mock;
 
-import gift.api.category.Category;
+import gift.api.category.domain.Category;
 import gift.api.product.domain.Product;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
